fix: address review findings — hash mismatch, fail-closed, DRY
Fixes from Phase 1 subagent review:

CRITICAL: Use Buffer (raw bytes) instead of UTF-8 string for diff
hashing, so Node `createHash` matches shell `shasum -a 256` behavior
even when diffs contain non-UTF-8 bytes.

HIGH: Fail-closed when computeDiffHash() returns empty — previously
silently passed, defeating the proof-of-work mechanism.

HIGH: Make checklist-required check advisory (not blocking) since
human devs, Dependabot, and non-agent automation don't use the
checklist workflow.

MEDIUM: Extract shared getDiffStats/parseDiffStat/getMergeBase into
diff-utils.ts to eliminate duplication between validators.

MEDIUM: Update stale file header comment to document new marker format.
